+++ packages/vue/src/components/organisms/SfCollectedProduct/SfCollectedProduct.vue
@@ -3,7 +3,11 @@
     <div class="sf-collected-product__main">
       <div class="sf-collected-product__details">
         <slot name="title" v-bind="{ title }">
-          <div class="sf-collected-product__title">{{ title }}</div>
+          <div class="sf-collected-product__title-wraper">
+            <SfLink :link="link" class="sf-collected-product__title">
+              {{ title }}
+            </SfLink>
+          </div>
         </slot>
         <slot name="price" v-bind="{ specialPrice, regularPrice }">
           <SfPrice
@@ -32,7 +36,6 @@
         <div class="sf-collected-product__quantity-wrapper">
           <SfQuantitySelector
             :qty="qty"
-            aria-label="Quantity"
             class="sf-collected-product__quantity-selector"
             @input="$emit('input', $event)"
           />
@@ -68,6 +71,7 @@ import SfImage from "../../atoms/SfImage/SfImage.vue";
 import SfCircleIcon from "../../atoms/SfCircleIcon/SfCircleIcon.vue";
 import SfButton from "../../atoms/SfButton/SfButton.vue";
 import SfQuantitySelector from "../../atoms/SfQuantitySelector/SfQuantitySelector.vue";
+import SfLink from "../../atoms/SfLink/SfLink.vue";
 export default {
   name: "SfCollectedProduct",
   components: {
@@ -77,6 +81,7 @@ export default {
     SfCircleIcon,
     SfPrice,
     SfQuantitySelector,
+    SfLink,
   },
   model: {
     prop: "qty",
@@ -131,6 +136,13 @@ export default {
     qty: {
       type: [Number, String],
       default: 1,
+    },
+    /**
+     * Link to product
+     */
+    link: {
+      type: [String, Object],
+      default: "",
     },
   },
   methods: {